% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download_url.R
\name{get_download_url}
\alias{get_download_url}
\title{Build the API URLs for each Tournament, to get games and players}
\usage{
get_download_url()
}
\value{
A character vector of API calls for each tournament
}
\description{
Build the API URLs for each Tournament, to get games and players
}
