% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournaments.R
\name{get_tournaments}
\alias{get_tournaments}
\title{Get list of Titled Tuesday codes from the Tournaments page \url{https://www.chess.com/tournament/live/titled-tuesdays}}
\usage{
get_tournaments(max_page = 20)
}
\arguments{
\item{max_page}{Maximum number of pages to crawl through}
}
\value{
A character vector of tournament URLs
}
\description{
Get list of Titled Tuesday codes from the Tournaments page \url{https://www.chess.com/tournament/live/titled-tuesdays}
}
