% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_games.R
\name{player_games}
\alias{player_games}
\title{Download and parse all games played by one or many chess.com accounts}
\usage{
player_games(usernames, years = "all")
}
\arguments{
\item{usernames}{Vector of usernames from whom to download games}

\item{years}{Vector of years from which to download games}
}
\value{
A data frame with one row per game, including the result, players, ratings, and PGN
}
\description{
Download and parse all games played by one or many chess.com accounts
}
\examples{
\dontrun{
# to download all of Hikaru's games since 2023
dat <- player_games(years = 2023:2024, usernames = "hikaru")

# to download all games by other players
dat <- player_games(usernames = c("jospem", "vladimirkramnik"))
}
}
