% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titled.tuesday-package.R
\docType{package}
\name{titled.tuesday-package}
\alias{titled.tuesday}
\alias{titled.tuesday-package}
\title{titled.tuesday: Load and parse Titled Tuesday data from the Chess.com API}
\description{
Download all PGNs and all standings for every Titled Tuesday Tournament.
}
\author{
\strong{Maintainer}: Igor Veiga \email{igor.rilave@hotmail.com}

}
\keyword{internal}
