% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titled_players.R
\name{titled_players}
\alias{titled_players}
\title{Download the list of all titled usernames registered on chess.com, separated by chess title}
\usage{
titled_players()
}
\value{
A list of character vectors with usernames, one for each chess title
}
\description{
Download the list of all titled usernames registered on chess.com, separated by chess title
}
\examples{
\dontrun{
usernames <- titled_players()
}
}
