% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titled_tuesday.R
\name{titled_tuesday}
\alias{titled_tuesday}
\title{Download and parse Titled Tuesday data}
\usage{
titled_tuesday()
}
\value{
A list with two data frames. One ($games) contains all games, the other ($players) contains the players and standings of each tournament.
}
\description{
Download and parse Titled Tuesday data
}
\examples{
\dontrun{
dat <- titled_tuesday()
}
}
