% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_games.R
\name{tt_games}
\alias{tt_games}
\title{Download and parse all Titled Tuesday games}
\usage{
tt_games(years = "all")
}
\arguments{
\item{years}{Vector of years from which to download games}
}
\value{
A data frame with one row per game
}
\description{
Download and parse all Titled Tuesday games
}
\examples{
\dontrun{
# download all TT games since 2023
tt <- tt_games(years = c(2023, 2024))
}
}
